LPS-71164 Merge checks to increase performance and so we don't have to keep all contents of all *.jsp in memory three times
